+++ curator/exceptions.py
@@ -43,3 +43,13 @@ class ActionTimeout(CuratorException):
     """
     Exception raised when an action fails to complete in the allotted time
     """
+
+class FailedSnapshot(CuratorException):
+    """
+    Exception raised when a snapshot does not complete with state SUCCESS
+    """
+
+class FailedRestore(CuratorException):
+    """
+    Exception raised when a Snapshot Restore does not restore all selected indices
+    """